Replaced context with prop object to remain simple
